requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
binaryornot==0.4.4
black==24.10.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-pkg==0.14
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
colorlog==6.9.0
cookiecutter==2.6.0
dj-database-url==3.0.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-modeltranslation==0.19.11
django-unfold==0.34.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
environ==1.0
environs==11.2.1
frozenlist==1.5.0
g4f==0.3.9.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==5.13.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jst-aicommit==1.3
jst-django==4.0.8
jst-django-core==1.1.1
kombu==5.4.2
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==24.1
pathspec==0.12.1
payme-pkg==3.0.24
pillow==11.0.0
platformdirs==4.3.6
polib==1.2.0
prompt-toolkit==3.0.36
propcache==0.2.1
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pyTelegramBotAPI==4.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
questionary==2.0.1
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.22.3
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.3
tenacity==9.0.0
text-unidecode==1.3
tqdm==4.67.1
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.18.3
zipfile36==0.1.3